app/helpers/group_helper.rb   - is_adm? and can_edit? methods -- the first could be "promoted" to the     application helper
app/models/group.rb
  - INTERFACE CHANGED: Group.include? now requires an "object array" as a group
    descriptor -- more cluttered but less "surprising"
  - many other internal changes

app/controllers/group_controller.rb
  - implemented all the basic functions

app/controllers/file_controller.rb
  - changed is_adm? to use the new Group.include? interface

app/views/group/list.rhtml
app/views/group/ajax_add_group.rjs
app/views/group/ajax_add_user.rjs
app/views/group/_list_element.rhtml
app/views/group/_group_element.rhtml
app/views/group/ajax_remove_user.rjs
app/views/group/ajax_delete_group.rjs
  - all these views implement the group management interface


git-svn-id: file:///var/lib/svn/shackboard/shackboard/trunk@92 6c9d4111-2920-0410-ac48-c0788669fd8e

+++ app/models/group.rb
@@ -2,26 +2,28 @@ class Group < ActiveRecord::Base
   has_many :group_memberships
   has_many :users, :through => :group_memberships
   def associate!(user) # {{{
-    raise ArgumentError, 'Argument is not a User' unless user.is_a? User
+    arr = object_to_array(user)
+    return false unless arr[0] == 'User'
     m = GroupMembership.find(:first,
-      :conditions => ['user_id = ? AND group_id = ?',
-  user.id, self.id]
+      :conditions => ['user_id = ? AND group_id = ?', arr[1], self.id]
     )
-    m = GroupMembership.new unless m
+    return false if m
+    m = GroupMembership.new
     m.associate(self, user)
     m.save
+    m
   end # }}}
   def remove!(user) # {{{
-    raise ArgumentError, 'Argument is not a User' unless user.is_a? User
+    arr = object_to_array(user)
+    return false unless arr[0] == 'User'
     m = GroupMembership.find(:first,
-      :conditions => ['user_id = ? AND group_id = ?',
-  user.id, self.id]
+      :conditions => ['user_id = ? AND group_id = ?', arr[1], self.id]
     )
     return false unless m
     m.destroy
   end # }}}
   def include?(user) # {{{
-    arr = arg_to_array(user)
+    arr = object_to_array(user)
     raise TypeError, "Argument is not a User or a ['User', user.id] array" \
       unless arr[0] == 'User'
     case arr[1].class.to_s
@@ -38,24 +40,71 @@ class Group < ActiveRecord::Base
     return true if m.is_a? GroupMembership
     return false
   end # }}}
-  def Group.include?(groupname, user) # {{{
-    g = Group.find_by_name(groupname)
+  def Group.include?(group, user) # {{{
+    g = Group.new.send(:object_to_object, group)
     return false unless g.is_a? Group
     g.include?(user)
   end # }}}
+  def Group.associate!(group, user) # {{{
+    g = Group.new.send(:object_to_object, group)
+    g.associate!(user)
+  end # }}}
+  def Group.remove!(group, user) # {{{
+    g = Group.new.send(:object_to_object, group)
+    g.remove!(user)
+  end # }}}
+  def can_edit?(user) # {{{
+    acl.can_edit?(user)
+  end # }}}
+  def can_edit(user) # {{{
+    acl.can_edit(user)
+  end # }}}
+  def remove_can_edit(user) # {{{
+    acl.remove_can_edit(user)
+  end # }}}
+  def acl_save # {{{
+    acl.save
+  end # }}}
+  private
   def acl # {{{
     return @acl if @acl
     @acl = AclMapping.map(self)
     return @acl if @acl
     @acl = Acl.new.attach_to(self)
   end # }}}
-  private
-  def arg_to_array(arg) # {{{
-    case arg.class.to_s
+  def object_to_array(obj) # {{{
+    case obj.class.to_s
     when 'Array'
-      arr = arg
+      arr = obj
     else
-      arr = [ arg.class.to_s, arg.id ]
+      arr = [ obj.class.to_s, obj.id ]
     end
+  end # }}}
+  def object_to_object(obj) # {{{
+    case obj.class.to_s
+    when 'Array'
+      case obj[1].class.to_s
+      when 'Fixnum'
+        return User.find(obj[1])  if obj[0] == 'User'
+        return Group.find(obj[1]) if obj[0] == 'Group'
+      when 'String'
+        return User.find_by_username(obj[1]) if obj[0] == 'User'
+        return Group.find_by_name(obj[1])    if obj[0] == 'Group'
+      end
+    when ('Group' or 'User')
+      return obj
+    end
+    nil
+  end # }}}
+  def arg_to_group(arg) # {{{
+    case arg.class.to_s
+    when 'Fixnum'
+      return Group.find(arg)
+    when 'String'
+      return Group.find_by_name(arg)
+    when 'Group'
+      return arg
+    end
+    nil
   end # }}}
 end
